+++ CHANGELOG.md
@@ -8,15 +8,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ```
+2021-04-24 23:40:58 Changed: Redesigned visit queue.
 2021-04-27 21:12:52 Fixed: Use `quantized_image` to detect changes and revisit pixels.
 2021-04-24 23:40:58 Changed: Redesigned for loops.
 2021-04-27 20:57:11 Fixed: Relocate palette construction to start of level loop.
 2021-04-27 17:26:05 Removed: Coarse levels/zooming, possible because of octree palette prediction.
 ```
 
-NOTE: Each of the above are major changes that effect the outcome of the annealing.
+NOTE: Each of the above are minor changes that effect the outcome of the annealing.
 
-## 2021-04-27 17:18:54 Version 0.1.0
+## 2021-04-27 17:18:54 [Version 0.1.0]
 
 ```
 2021-04-25 18:59:48 Fixed: shortcut to set identity matrix triggers a compiler bug.
@@ -27,3 +28,4 @@ NOTE: Each of the above are major changes that effect the outcome of the anneali
 ```
 
 [Unreleased]: https://github.com/xyzzy/scq/compare/v0.1.0...HEAD
+[Version 0.1.0]: https://github.com/xyzzy/scq/tree/v0.1.0